requirements.txt
numpy
pillow
tqdm
wandb
torch
torchvision
